requirements.txt
pypdf
pypdf2
swibots
openai
langchain
chromadb
python-decouple